# team-service/swagger.yaml
swagger: '2.0'

info:
  version: "v1.0"
  title: Teams Info Service

definitions:
  InfrastructureAccount:
    type: object
    properties:
      id:
        type: string
        example: 2892853535
      type:
        type: string
        example: aws
  ShortTeamInfo:
    type: object
    properties:
      id:
        type: string
        example: stups
      name:
        type: string
        example: stups
      infrastructure-accounts:
        type: array
        items:
          $ref: "#/definitions/InfrastructureAccount"
  FullTeamInfo:
    type: object
    properties:
      id:
        type: string
        example: stups
      name:
        type: string
        example: stups
      description:
        type: string
        example: Cloud Engineers
      dn:
        type: string
        example: ou=users,ou=apps
      infrastructure-accounts:
        type: array
        items:
          $ref: "#/definitions/InfrastructureAccount"
      members:
        type: array
        items:
          type: string
          example: npiccolotto

paths:
  # This is a path endpoint. Change it.
  /teams:
    get:
      description: Get a list of teams

      # Expected responses for this operation:
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: "#/definitions/ShortTeamInfo"
  /teams/{team_id}:
    get:
      description: detailed view for a team
      parameters:
      - name: team_id
        in: path
        type: string
        description: the team's id
        required: true

      responses:
        200:
          description: Successful response
          schema:
            $ref: "#/definitions/FullTeamInfo"
  /user/{user_id}:
    get:
      description: list of teams a user is in
      parameters:
      - name: user_id
        in: path
        type: string
        description: the user id
        required: true

      responses:
        200:
          description: Successful response
          schema:
            type: array
            items:
              $ref: "#/definitions/ShortTeamInfo"